[INLONG-4632][Manager] Add the check method back.
