+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/sink/AbstractSinkOperator.java
@@ -21,12 +21,15 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.inlong.manager.common.enums.ErrorCodeEnum;
+import org.apache.inlong.manager.common.enums.FieldType;
 import org.apache.inlong.manager.common.enums.GlobalConstants;
 import org.apache.inlong.manager.common.enums.SinkStatus;
+import org.apache.inlong.manager.common.enums.SinkType;
 import org.apache.inlong.manager.common.exceptions.BusinessException;
 import org.apache.inlong.manager.common.pojo.sink.SinkField;
 import org.apache.inlong.manager.common.pojo.sink.SinkRequest;
 import org.apache.inlong.manager.common.pojo.sink.StreamSink;
+import org.apache.inlong.manager.common.pojo.sink.iceberg.IcebergColumnInfo;
 import org.apache.inlong.manager.common.util.CommonBeanUtils;
 import org.apache.inlong.manager.common.util.Preconditions;
 import org.apache.inlong.manager.dao.entity.StreamSinkEntity;
@@ -154,7 +157,11 @@ public abstract class AbstractSinkOperator implements StreamSinkOperation {
         String streamId = request.getInlongStreamId();
         String sinkType = request.getSinkType();
         Integer sinkId = request.getId();
+        boolean checkField = StringUtils.equals(sinkType, SinkType.SINK_ICEBERG);
         for (SinkField fieldInfo : fieldList) {
+            if (checkField) {
+                checkFieldInfo(fieldInfo);
+            }
             StreamSinkFieldEntity fieldEntity = CommonBeanUtils.copyProperties(fieldInfo, StreamSinkFieldEntity::new);
             if (StringUtils.isEmpty(fieldEntity.getFieldComment())) {
                 fieldEntity.setFieldComment(fieldEntity.getFieldName());
@@ -169,6 +176,26 @@ public abstract class AbstractSinkOperator implements StreamSinkOperation {
 
         sinkFieldMapper.insertAll(entityList);
         LOGGER.info("success to save field");
+    }
+
+    private void checkFieldInfo(SinkField field) {
+        if (FieldType.forName(field.getFieldType()) == FieldType.DECIMAL) {
+            IcebergColumnInfo info = IcebergColumnInfo.getFromJson(field.getExtParams());
+            if (info.getPrecision() == null || info.getScale() == null) {
+                String errorMsg = String.format("precision or scale not specified for decimal field (%s)",
+                        field.getFieldName());
+                LOGGER.error("field info check error: {}", errorMsg);
+                throw new BusinessException(errorMsg);
+            }
+            if (info.getPrecision() < info.getScale()) {
+                String errorMsg = String.format(
+                        "precision (%d) must be greater or equal than scale (%d) for decimal field (%s)",
+                        info.getPrecision(), info.getScale(), field.getFieldName());
+                LOGGER.error("field info check error: {}", errorMsg);
+                throw new BusinessException(errorMsg);
+            }
+        }
+        return;
     }
 
 }
