+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/sink/es/ElasticsearchSinkOperation.java
@@ -116,7 +116,7 @@ public class ElasticsearchSinkOperation extends AbstractSinkOperator {
             ElasticsearchSinkDTO dto = ElasticsearchSinkDTO.getFromRequest(sinkRequest);
             targetEntity.setExtParams(objectMapper.writeValueAsString(dto));
         } catch (Exception e) {
-            LOGGER.error("Error occurred while saving or updating sink info", e);
+            LOGGER.error("parsing json string to sink info failed", e);
             throw new BusinessException(ErrorCodeEnum.SINK_SAVE_FAILED.getMessage());
         }
     }
