+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/sink/hbase/HBaseSinkOperation.java
@@ -115,7 +115,7 @@ public class HBaseSinkOperation extends AbstractSinkOperator {
             HBaseSinkDTO dto = HBaseSinkDTO.getFromRequest(sinkRequest);
             targetEntity.setExtParams(objectMapper.writeValueAsString(dto));
         } catch (Exception e) {
-            LOGGER.error("Error occurred while saving or updating sink info", e);
+            LOGGER.error("parsing json string to sink info failed", e);
             throw new BusinessException(ErrorCodeEnum.SINK_SAVE_FAILED.getMessage());
         }
     }
