+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/sink/hdfs/HdfsSinkOperation.java
@@ -113,7 +113,7 @@ public class HdfsSinkOperation extends AbstractSinkOperator {
             HdfsSinkDTO dto = HdfsSinkDTO.getFromRequest(sinkRequest);
             targetEntity.setExtParams(objectMapper.writeValueAsString(dto));
         } catch (Exception e) {
-            LOGGER.error("Error occurred while saving or updating sink info", e);
+            LOGGER.error("parsing json string to sink info failed", e);
             throw new BusinessException(ErrorCodeEnum.SINK_SAVE_FAILED.getMessage());
         }
     }
