+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/sink/iceberg/IcebergSinkOperation.java
@@ -22,14 +22,12 @@ import com.github.pagehelper.Page;
 import com.github.pagehelper.PageInfo;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.inlong.manager.common.enums.ErrorCodeEnum;
-import org.apache.inlong.manager.common.enums.FieldType;
 import org.apache.inlong.manager.common.enums.SinkType;
 import org.apache.inlong.manager.common.exceptions.BusinessException;
 import org.apache.inlong.manager.common.pojo.sink.SinkField;
 import org.apache.inlong.manager.common.pojo.sink.SinkListResponse;
 import org.apache.inlong.manager.common.pojo.sink.SinkRequest;
 import org.apache.inlong.manager.common.pojo.sink.StreamSink;
-import org.apache.inlong.manager.common.pojo.sink.iceberg.IcebergColumnInfo;
 import org.apache.inlong.manager.common.pojo.sink.iceberg.IcebergSink;
 import org.apache.inlong.manager.common.pojo.sink.iceberg.IcebergSinkDTO;
 import org.apache.inlong.manager.common.pojo.sink.iceberg.IcebergSinkListResponse;
@@ -64,26 +62,6 @@ public class IcebergSinkOperation extends AbstractSinkOperator {
     @Override
     public Boolean accept(SinkType sinkType) {
         return SinkType.ICEBERG.equals(sinkType);
-    }
-
-    private void checkFieldInfo(SinkField field) {
-        if (FieldType.forName(field.getFieldType()) == FieldType.DECIMAL) {
-            IcebergColumnInfo info = IcebergColumnInfo.getFromJson(field.getExtParams());
-            if (info.getPrecision() == null || info.getScale() == null) {
-                String errorMsg = String.format("precision or scale not specified for decimal field (%s)",
-                        field.getFieldName());
-                LOGGER.error("field info check error: {}", errorMsg);
-                throw new BusinessException(errorMsg);
-            }
-            if (info.getPrecision() < info.getScale()) {
-                String errorMsg = String.format(
-                        "precision (%d) must be greater or equal than scale (%d) for decimal field (%s)",
-                        info.getPrecision(), info.getScale(), field.getFieldName());
-                LOGGER.error("field info check error: {}", errorMsg);
-                throw new BusinessException(errorMsg);
-            }
-        }
-        return;
     }
 
     @Override
@@ -137,7 +115,7 @@ public class IcebergSinkOperation extends AbstractSinkOperator {
             IcebergSinkDTO dto = IcebergSinkDTO.getFromRequest(sinkRequest);
             targetEntity.setExtParams(objectMapper.writeValueAsString(dto));
         } catch (Exception e) {
-            LOGGER.error("Error occurred while saving or updating sink info", e);
+            LOGGER.error("parsing json string to sink info failed", e);
             throw new BusinessException(ErrorCodeEnum.SINK_SAVE_FAILED.getMessage());
         }
     }
