+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/sink/sqlserver/SqlServerSinkOperation.java
@@ -112,7 +112,7 @@ public class SqlServerSinkOperation extends AbstractSinkOperator {
             SqlServerSinkDTO dto = SqlServerSinkDTO.getFromRequest(sinkRequest);
             targetEntity.setExtParams(objectMapper.writeValueAsString(dto));
         } catch (Exception e) {
-            LOGGER.error("Error occurred while saving or updating sink info", e);
+            LOGGER.error("parsing json string to sink info failed", e);
             throw new BusinessException(ErrorCodeEnum.SINK_SAVE_FAILED.getMessage());
         }
     }
