+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/sink/tdsqlpostgresql/TDSQLPostgreSQLSinkOperation.java
@@ -112,7 +112,7 @@ public class TDSQLPostgreSQLSinkOperation extends AbstractSinkOperator {
             TDSQLPostgreSQLSinkDTO dto = TDSQLPostgreSQLSinkDTO.getFromRequest(sinkRequest);
             targetEntity.setExtParams(objectMapper.writeValueAsString(dto));
         } catch (Exception e) {
-            LOGGER.error("Error occurred while saving or updating sink info", e);
+            LOGGER.error("parsing json string to sink info failed", e);
             throw new BusinessException(ErrorCodeEnum.SINK_SAVE_FAILED.getMessage());
         }
     }
